Using River, crawler started
River is a queueing system built on Postgres.

It is a bit early days, but seems to do what is needed, and I like that
it lives on PG as opposed to a more ephemeral engine.

This is the start of an abstracted interface for building clients that
interact with the queue and then communicate via channels to the rest of
an application.